Add SVG for Mastodon accounts (#72)
Mastodon is a federated microblogging platform (just like Twitter). The
main difference is that we have different instances, managed by
different people and we are able to follow users from all instances in a
very decentralized way. Adding the logo of Mastodon will help users to
add it as another option for "socialOptions" config.